requirements.txt
google-ai-generativelanguage>=0.4.0
